Fixed input height and width

+++ horn.h
@@ -7,7 +7,7 @@
 class Horn {
 public:
   Horn();
-  void setup(Ui::MainWindow *, uint8_t, uint8_t);
+  void setup(Ui::MainWindow *, uint16_t, uint16_t);
   void horn_on();
   void horn_off();
   ~Horn(){};
